raspberry_pi/requirements.txt
pygame==2.1.2
pyserial==3.5 